% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cbtext-package.R
\docType{package}
\name{cbtext-package}
\alias{cbtext}
\alias{cbtext-package}
\title{cbtext: scrape central banks}
\description{
This package scrape the internet to get central bank texts into R.
}
\author{
\strong{Maintainer}: Martin Baumgärtner \email{martin.baumgaertner@posteo.de} (www.martinbaumgaertner.de)

}
\keyword{internal}
