% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tibble.R
\name{create_tibble}
\alias{create_tibble}
\title{create_tibble}
\usage{
create_tibble(names, rlength)
}
\arguments{
\item{names}{columnnames}

\item{rlength}{rows}
}
\value{
tibble
}
\description{
prepares and empty tibble
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[dplyr]{reexports}}
}
