% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_features.R
\name{doc_features}
\alias{doc_features}
\title{doc_features}
\usage{
doc_features()
}
\value{
OUTPUT_DESCRIPTION
}
\description{
returns document features
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
