% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bis_links.R
\name{read_bis_links}
\alias{read_bis_links}
\title{Hauptfunktion zum Lesen von XML-Links}
\usage{
read_bis_links(
  urls = c("https://www.bis.org/sitemap.xml?documents=2005,2006,2007,2008,2009",
    "https://www.bis.org/sitemap.xml?documents=2010,2011,2012,2013,2014",
    "https://www.bis.org/sitemap.xml?documents=2015,2016,2017,2018,2019",
    "https://www.bis.org/sitemap.xml?documents=2020,2021,2022,2023")
)
}
\arguments{
\item{urls}{Eine Vektor von Zeichenketten, die die URLs der zu lesenden XML-Dateien darstellen.}
}
\value{
Ein tibble, das alle Links aus den XML-Dateien enthält, die "/review/" enthalten, klassifiziert nach ihrem Typ.
}
\description{
Diese Funktion liest XML-Dateien von einer Liste von URLs und konvertiert sie in ein tibble.
Sie filtert dann nach Links, die "/review/" enthalten und klassifiziert sie nach ihrem Typ (PDF oder HTML).
}
\examples{
\dontrun{
read_bis_links(c("https://www.bis.org/sitemap.xml?documents=2005,2006,2007,2008,2009"))
}
}
