% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bis_links.R
\name{read_xml_link}
\alias{read_xml_link}
\title{Funktion zum Lesen und Umwandeln einer XML-Datei}
\usage{
read_xml_link(url)
}
\arguments{
\item{url}{Eine Zeichenkette, die die URL der zu lesenden XML-Datei darstellt.}
}
\value{
Ein tibble, das aus der XML-Datei erstellt wurde. Bei Fehlern beim Lesen der Datei gibt die Funktion NULL zurück.
}
\description{
Diese Funktion liest eine XML-Datei von einer gegebenen URL und konvertiert sie in ein tibble.
}
\examples{
\dontrun{
read_xml_link("https://www.bis.org/sitemap.xml?documents=2005,2006,2007,2008,2009")
}
}
