% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_source.R
\name{scrape_source}
\alias{scrape_source}
\title{scrape_source}
\usage{
scrape_source(source, parts, spacy = T)
}
\arguments{
\item{source}{character vector describing central banks}

\item{parts}{character vector describing parts to scrape}
}
\value{
tibble
}
\description{
Scrape data from central bank source
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
  x<-scrape_source(list(name="boe"),c("minutes"))
  x<-scrape_source(list(name="ecb"),c("minutes","pc","economic_outlook","blog"))
  x<-scrape_source(list(name="fed"),c("minutes","pc","red","transkript","agenda",
  "blue","green_sub","teala","tealb","green1","green2","beige"))
  x<-scrape_source(list(name="boj"),c("minutes","economic_report","release","outlook_report"))
  x<-scrape_source(list(name="poland"),c("minutes","inflation_report","release"))
  x<-scrape_source(list(name="iceland"),c("minutes"))
  x<-scrape_source(list(name="australia"),c("minutes"))
  x<-scrape_source(list(name="riksbank"),c("minutes","economicreview","mpreports"))
 }
}
}
